YWolfeee: change setting reading method to get prepared for convenient test
