+++ Appointment/utils/scheduler_func.py
@@ -1,6 +1,6 @@
 # YWolfeee:
 # 本py文件保留所有需要与scheduler交互的函数。
-from Appointment import global_info
+from YPUnderground import global_info
 from apscheduler.schedulers.background import BackgroundScheduler
 from django_apscheduler.jobstores import DjangoJobStore, register_events, register_job
